requirements.txt
# requirements.txt
openai
dotenv
pytest